requirements.txt
tkcalendar
psycopg2-binary
plyer